Refactor cases pipelines.
