+++ dags/corona_cases/cases/__init__.py
@@ -1,0 +1,3 @@
+from .download_corona_cases import covid_cases_etl, CASES_ARGS
+
+__all__ = ["covid_cases_etl", "CASES_ARGS"]
