+++ dags/corona_cases/cases/download_corona_cases.py
@@ -11,10 +11,15 @@ from dags.helpers.dag_helpers import connect_to_db_and_insert
 from dags.helpers.test_helpers import set_env_variable_from_dag_config_if_present
 
 
+URL = "https://prod-hub-indexer.s3.amazonaws.com/files/dd4580c810204019a7b8eb3e0b329dd6/0/full/4326/dd4580c810204019a7b8eb3e0b329dd6_0_full_4326.csv"  # noqa: E501
+SCHEMA = "coronacases"
+TABLE = "german_counties_more_info"
+
+CASES_ARGS = [URL, SCHEMA, TABLE]
+
+
 @curry
-def download_csv_and_upload_to_postgres(
-    url: str, schema: str, table: str, **kwargs
-) -> DBContext:
+def covid_cases_etl(url: str, schema: str, table: str, **kwargs) -> DBContext:
     return R.pipe(
         set_env_variable_from_dag_config_if_present("TARGET_DB"),
         lambda *args: download_csv(url),
