+++ dags/corona_cases/test_integration_dag.py
@@ -3,14 +3,14 @@ from dags.database.execute_sql import query_all_elements
 from airflow.models import DagBag
 
 from dags.helpers.test_helpers import execute_dag
-from dags.corona_cases.corona_cases_dag import SCHEMA, TABLE
+from dags.corona_cases.dag import SCHEMA, TABLE
 
 URL = "http://static-files/static/coronacases.csv"
 
 
 def test_dag_loads_with_no_errors():
     dag_bag = DagBag(include_examples=False)
-    dag_bag.process_file("corona_cases_dag.py")
+    dag_bag.process_file("dag.py")
     assert len(dag_bag.import_errors) == 0
